Move json-ld attributes into openannotation.py
